requirements.txt
djangorestframework==3.16.1
django-simple-captcha==0.6.2
fastapi==0.118.3
pydantic==2.9.2
SQLAlchemy==2.0.44
python-dotenv==1.1.1
tqdm==4.67.1
pysrt==1.1.2
librosa==0.11.0
soundfile==0.13.1
edge-tts==7.2.3
uvicorn[standard]
TTS==0.22.0
transformers==4.39.3
modelscope==1.17.1
faster-whisper==1.2.0
demucs==4.0.1
azure-core==1.36.0
azure-cognitiveservices-speech==1.46.0